+++ tests/Gedmo/Sluggable/Fixture/Handler/TreeSlug.php
@@ -24,6 +24,7 @@ use Gedmo\Tree\Entity\Repository\NestedTreeRepository;
  * @ORM\Entity(repositoryClass="Gedmo\Tree\Entity\Repository\NestedTreeRepository")
  */
 #[ORM\Entity(repositoryClass: NestedTreeRepository::class)]
+#[Gedmo\Tree(type: 'nested')]
 class TreeSlug
 {
     /**
@@ -71,6 +72,7 @@ class TreeSlug
      */
     #[ORM\ManyToOne(targetEntity: self::class)]
     #[ORM\JoinColumn(name: 'parent_id', referencedColumnName: 'id', onDelete: 'CASCADE')]
+    #[Gedmo\TreeParent]
     private $parent;
 
     /**
@@ -85,6 +87,7 @@ class TreeSlug
      * @ORM\Column(type="integer")
      */
     #[ORM\Column(type: Types::INTEGER)]
+    #[Gedmo\TreeLeft]
     private $lft;
 
     /**
@@ -94,6 +97,7 @@ class TreeSlug
      * @ORM\Column(type="integer")
      */
     #[ORM\Column(type: Types::INTEGER)]
+    #[Gedmo\TreeRight]
     private $rgt;
 
     /**
@@ -103,6 +107,7 @@ class TreeSlug
      * @ORM\Column(type="integer")
      */
     #[ORM\Column(type: Types::INTEGER)]
+    #[Gedmo\TreeRoot]
     private $root;
 
     /**
@@ -112,6 +117,7 @@ class TreeSlug
      * @ORM\Column(name="lvl", type="integer")
      */
     #[ORM\Column(name: 'lvl', type: Types::INTEGER)]
+    #[Gedmo\TreeLevel]
     private $level;
 
     public function __construct()