+++ CHANGELOG.md
@@ -26,6 +26,7 @@ a release.
 - Sortable: Support to use annotations as attributes on PHP >= 8.0.
 - Sluggable: Support to use annotations as attributes on PHP >= 8.0.
 - Uploadable: Support to use annotations as attributes on PHP >= 8.0.
+- Tree: Support to use annotations as attributes on PHP >= 8.0.
 
 ### Fixed
 - Blameable, IpTraceable, Timestampable: Type handling for the tracked field values configured in the origin field.